Super refactor

+++ Backend/src/main/java/fpt/edu/vn/Backend/service/CurrencyService.java
@@ -2,10 +2,11 @@ package fpt.edu.vn.Backend.service;
 
 import fpt.edu.vn.Backend.pojo.CurrencyType;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface CurrencyService {
-    Map<CurrencyType, Double> getExchangeRates();
+    Map<CurrencyType, BigDecimal> getExchangeRates();
     Double getExchangeRate(CurrencyType currency);
     long getLastUpdate();
 }
